Add Lumen psalm plugin
